+++ internal/client/hearts/dto_test.go
@@ -1,4 +1,4 @@
-package hearts
+package heartsclient
 
 import (
 	"encoding/json"